Refactor server code for testing
